backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database — SQLite (no external server needed)
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilities
httpx==0.28.1
tenacity==9.0.0

# Simulation / Testing
websockets==14.1
